Handle EEXIST on registration
Currently we've got a bit of a conundrum:

* When curl wants us to remove a socket, the socket may or may not be
  open.
* When curl wants us to register a socket, it may or may not have been
  previously registered

To solve the first problem we can't call deregister for fear of
deregistering a different file descriptor. That means that to solve the
second problem we may get `EEXIST` when attempting to register a socket
and we need to handle that.

This commit maps `EEXIST` to `reregister` during registration to handle
this case.